Refactoring packages in graph
