+++ android_client/presentation/src/main/java/smarthome/client/presentation/scripts/addition/controllers/epoxy/ControllerView.kt
@@ -13,7 +13,7 @@ import com.airbnb.epoxy.ModelView
 import com.airbnb.epoxy.TextProp
 import kotlinx.android.synthetic.main.scripts_controller_item.view.*
 import smarthome.client.presentation.R
-import smarthome.client.presentation.scripts.addition.graph.Position
+import smarthome.client.presentation.util.Position
 import smarthome.client.presentation.scripts.addition.graph.events.drag.GraphDragEvent
 import smarthome.client.presentation.util.CustomDragShadowBuilder
 import smarthome.client.presentation.util.inflate
@@ -30,7 +30,9 @@ class ControllerView @JvmOverloads constructor(
         inflate(R.layout.scripts_controller_item)
         val detector = GestureDetector(context, object : GestureDetector.SimpleOnGestureListener() {
             override fun onLongPress(pressEvent: MotionEvent) {
-                val info = onDragStarted?.invoke(Position(pressEvent.x, pressEvent.y))
+                val info = onDragStarted?.invoke(
+                    Position(pressEvent.x,
+                        pressEvent.y))
 
                 val data = ClipData.newPlainText("sad", "asdf")
                 val shadowBuilder = CustomDragShadowBuilder(this@ControllerView, pressEvent)
