--- android_client/presentation/src/main/java/smarthome/client/presentation/scripts/addition/graph/GraphBlockIdentifier.kt
@@ -1,3 +1,0 @@
-package smarthome.client.presentation.scripts.addition.graph
-
-interface GraphBlockIdentifier
