--- android_client/presentation/src/main/java/smarthome/client/presentation/scripts/addition/graph/GraphBlockView.kt
@@ -1,7 +1,0 @@
-package smarthome.client.presentation.scripts.addition.graph
-
-import smarthome.client.presentation.scripts.addition.graph.views.state.GraphBlock
-
-interface GraphBlockView {
-    fun setData(block: GraphBlock)
-}
