+++ android_client/presentation/src/main/java/smarthome/client/presentation/scripts/addition/graph/blockviews/GraphBlockView.kt
@@ -1,0 +1,7 @@
+package smarthome.client.presentation.scripts.addition.graph.blockviews
+
+import smarthome.client.presentation.scripts.addition.graph.blockviews.state.GraphBlock
+
+interface GraphBlockView {
+    fun setData(block: GraphBlock)
+}
