+++ android_client/presentation/src/main/java/smarthome/client/presentation/scripts/addition/graph/blockviews/factory/BlockFactoryConstants.kt
@@ -1,0 +1,3 @@
+package smarthome.client.presentation.scripts.addition.graph.blockviews.factory
+
+const val CONTROLLER_FACTORY = "CONTROLLER_FACTORY"
