+++ android_client/presentation/src/main/java/smarthome/client/presentation/scripts/addition/graph/blockviews/factory/GraphBlockFactoryResolver.kt
@@ -1,0 +1,7 @@
+package smarthome.client.presentation.scripts.addition.graph.blockviews.factory
+
+import smarthome.client.presentation.scripts.addition.graph.blockviews.state.GraphBlock
+
+interface GraphBlockFactoryResolver {
+    fun resolve(block: GraphBlock): GraphBlockFactory
+}
