+++ android_client/presentation/src/main/java/smarthome/client/presentation/scripts/addition/graph/blockviews/state/GraphBlockResolver.kt
@@ -1,6 +1,6 @@
-package smarthome.client.presentation.scripts.addition.graph.views.state
+package smarthome.client.presentation.scripts.addition.graph.blockviews.state
 
-import smarthome.client.presentation.scripts.addition.graph.GraphBlockIdentifier
+import smarthome.client.presentation.scripts.addition.graph.identifier.GraphBlockIdentifier
 import smarthome.client.presentation.scripts.addition.graph.events.drag.GraphDragEvent
 
 interface GraphBlockResolver {
