+++ android_client/presentation/src/main/java/smarthome/client/presentation/scripts/addition/graph/blockviews/state/GraphBlockResolverImpl.kt
@@ -1,7 +1,7 @@
-package smarthome.client.presentation.scripts.addition.graph.views.state
+package smarthome.client.presentation.scripts.addition.graph.blockviews.state
 
-import smarthome.client.presentation.scripts.addition.graph.ControllerGraphBlockIdentifier
-import smarthome.client.presentation.scripts.addition.graph.GraphBlockIdentifier
+import smarthome.client.presentation.scripts.addition.graph.identifier.ControllerGraphBlockIdentifier
+import smarthome.client.presentation.scripts.addition.graph.identifier.GraphBlockIdentifier
 import smarthome.client.presentation.scripts.addition.graph.events.drag.ControllerDragEvent
 import smarthome.client.presentation.scripts.addition.graph.events.drag.GraphDragEvent
 
@@ -9,7 +9,8 @@ class GraphBlockResolverImpl : GraphBlockResolver {
 
     override fun resolveIdentifier(event: GraphDragEvent): GraphBlockIdentifier {
         return when (event) {
-            is ControllerDragEvent -> ControllerGraphBlockIdentifier(event.id)
+            is ControllerDragEvent -> ControllerGraphBlockIdentifier(
+                event.id)
             else -> throw IllegalStateException("No identifier associated for event $event")
         }
     }
@@ -17,7 +18,8 @@ class GraphBlockResolverImpl : GraphBlockResolver {
     override fun createBlock(event: GraphDragEvent): GraphBlock {
         return when (event) {
             is ControllerDragEvent -> ControllerBlock(
-                ControllerGraphBlockIdentifier(event.id),
+                ControllerGraphBlockIdentifier(
+                    event.id),
                 event.dragInfo.position
             )
             else -> throw IllegalStateException("No graph block associated for event $event")
