+++ android_client/presentation/src/main/java/smarthome/client/presentation/scripts/addition/graph/identifier/ControllerGraphBlockIdentifier.kt
@@ -1,4 +1,4 @@
-package smarthome.client.presentation.scripts.addition.graph
+package smarthome.client.presentation.scripts.addition.graph.identifier
 
 class ControllerGraphBlockIdentifier(val id: Long) :
     GraphBlockIdentifier {
