+++ android_client/presentation/src/main/java/smarthome/client/presentation/scripts/addition/graph/identifier/GraphBlockIdentifier.kt
@@ -1,0 +1,3 @@
+package smarthome.client.presentation.scripts.addition.graph.identifier
+
+interface GraphBlockIdentifier
