--- android_client/presentation/src/main/java/smarthome/client/presentation/scripts/addition/graph/views/factory/BlockFactoryConstants.kt
@@ -1,3 +1,0 @@
-package smarthome.client.presentation.scripts.addition.graph.views.factory
-
-const val CONTROLLER_FACTORY = "CONTROLLER_FACTORY"
