--- android_client/presentation/src/main/java/smarthome/client/presentation/scripts/addition/graph/views/factory/GraphBlockFactoryResolver.kt
@@ -1,7 +1,0 @@
-package smarthome.client.presentation.scripts.addition.graph.views.factory
-
-import smarthome.client.presentation.scripts.addition.graph.views.state.GraphBlock
-
-interface GraphBlockFactoryResolver {
-    fun resolve(block: GraphBlock): GraphBlockFactory
-}
