+++ android_client/presentation/src/test/java/smarthome/client/presentation/scripts/addition/controllers/ControllersViewViewModelTest.kt
@@ -18,10 +18,10 @@ import org.koin.dsl.module
 import smarthome.client.domain.api.conrollers.usecases.ObserveControllerUseCase
 import smarthome.client.domain.api.devices.usecase.GetGeneralDevicesInfo
 import smarthome.client.entity.Controller
-import smarthome.client.presentation.scripts.addition.graph.*
 import smarthome.client.presentation.scripts.addition.graph.events.GraphEvent
 import smarthome.client.presentation.scripts.addition.graph.events.GraphEventBus
 import smarthome.client.presentation.scripts.addition.graph.events.drag.*
+import smarthome.client.presentation.util.Position
 import smarthome.client.util.DataStatus
 
 class ControllersViewViewModelTest {
